requirements.txt
pushbullet.py
python-decouple
requests